Added full support for fetching and tweeting of longer tweets (#8)
With the update to Twitter that allows you to make 280 character tweets,
you have to specify "tweet_mode: extended" when getting JSON tweets,
otherwise you get truncated 140 character versions. This also changes
the key in the retrieved data to be "full_text" instead of just "text".
I added a check so that if "full_text" is present to define that as the
tweet text, and if not to use "text" (for backwards compatibility with
existing JSON archives).
Also updated the hard limits in places to be 280 instead of 140.